Catalog export to be performed always
We need to run the export job even when the job fails
so that the export files show the latest version of
the saved catalog.

Later when we sent the files to DW we will need to
only send them when the import is successful and/or
the user wants to send it.